# api/openapi.yaml
openapi: 3.0.3
info:
  title: otelfleet API
  description: >
    Control-plane REST API for the otelfleet platform. Browser sessions are
    cookie-based (`__Host-otelfleet_session`, HttpOnly). Mutating requests must
    send the `X-CSRF-Token` header obtained from `GET /api/v1/me`.
    OIDC browser flows live outside this spec at `/auth/{provider}/start` and
    `/auth/{provider}/callback`.
  version: 0.1.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /
...
paths:
...
  /api/v1/auth/dev-login:
    post:
      tags: [auth]
      operationId: devLogin
      summary: Password-less local development login (disabled in production)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "204": { description: Session cookie set }
        "403": { description: Dev login disabled, $ref: "#/components/responses/Forbidden" }
...
  /api/v1/users:
    get:
      tags: [admin]
      operationId: listUsers
      summary: List UI users (admin only)
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: object
                required: [users]
                properties:
                  users:
                    type: array
                    items: { $ref: "#/components/schemas/UserAccount" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [admin]
      operationId: inviteUser
      summary: Invite a user by email; the role applies on their first SSO login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email: { type: string, format: email }
                role: { $ref: "#/components/schemas/Role" }
                customerIds:
                  type: array
                  items: { type: string, format: uuid }
                  description: >-
                    Tenant-scope grants. When set, a non-admin user is limited
                    to these customers; omit or empty = access to all customers.
                    Ignored for admins.
      responses:
        "201":
          description: Invited user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserAccount" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { description: Email already exists, $ref: "#/components/responses/Conflict" }
...
  /api/v1/settings/auth-providers:
    get:
      tags: [admin]
      operationId: listAuthProviderConfigs
      summary: Configured SSO providers incl. env-defined ones (admin only, no secrets)
      responses:
        "200":
          description: Providers
          content:
            application/json:
              schema:
                type: object
                required: [providers]
                properties:
                  providers:
                    type: array
                    items: { $ref: "#/components/schemas/AuthProviderConfig" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [admin]
      operationId: createAuthProviderConfig
      summary: Add an SSO provider (client secret is encrypted at rest)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthProviderConfigCreate" }
      responses:
        "201":
          description: Created provider
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthProviderConfig" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { description: Provider name already exists, $ref: "#/components/responses/Conflict" }
...
  /api/v1/settings/api-tokens:
    get:
      tags: [admin]
      operationId: listApiTokens
      summary: List management-API tokens (admin only, secrets never returned)
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema:
                type: object
                required: [tokens]
                properties:
                  tokens:
                    type: array
                    items: { $ref: "#/components/schemas/ApiToken" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [admin]
      operationId: createApiToken
      summary: Create a management-API token (secret returned exactly once)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, role]
              properties:
                name: { type: string, minLength: 1, maxLength: 200 }
                role: { $ref: "#/components/schemas/Role" }
                expiresAt: { type: string, format: date-time, nullable: true, description: Defaults to no expiry }
      responses:
        "201":
          description: Token including the one-time secret
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiTokenCreated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/v1/audit:
    get:
      tags: [admin]
      operationId: listAuditLog
      summary: Query the audit log (newest first, cursor-paged)
      parameters:
        - { name: action, in: query, schema: { type: string } }
        - { name: entityType, in: query, schema: { type: string } }
        - { name: customerId, in: query, schema: { type: string, format: uuid } }
        - { name: actorUserId, in: query, schema: { type: string, format: uuid } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
        - { name: beforeId, in: query, schema: { type: integer, format: int64 }, description: Cursor from the previous page }
      responses:
        "200":
          description: Audit entries
          content:
            application/json:
              schema:
                type: object
                required: [entries]
                properties:
                  entries:
                    type: array
                    items: { $ref: "#/components/schemas/AuditEntry" }
                  nextBeforeId:
                    type: integer
                    format: int64
                    nullable: true
                    description: Pass as beforeId to fetch the next page; null when exhausted.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Insufficient role
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: not_found }
        message: { type: string }
    Role:
      type: string
      enum: [admin, operator, viewer]
...
    ApiToken:
      type: object
      required: [id, name, tokenPrefix, role, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        tokenPrefix: { type: string, example: otm_pat_a1b2c3d4 }
        role: { $ref: "#/components/schemas/Role" }
        createdBy: { type: string, nullable: true, description: Creator email }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time, nullable: true }
        lastUsedAt: { type: string, format: date-time, nullable: true }
        revokedAt: { type: string, format: date-time, nullable: true }
    ApiTokenCreated:
      allOf:
        - $ref: "#/components/schemas/ApiToken"
        - type: object
          required: [secret]
          properties:
            secret:
              type: string
              description: "Full token (otm_pat_…). Shown once; send as an Authorization: Bearer header."
    UserAccount:
      type: object
      required: [id, email, role, disabled, invited, identities, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string, nullable: true }
        role: { $ref: "#/components/schemas/Role" }
        disabled: { type: boolean }
        invited:
          type: boolean
          description: True until the user logs in for the first time (no linked identity yet).
        identities:
          type: array
          description: Linked login identities (provider names).
          items: { type: string }
        customerIds:
          type: array
          description: >-
            Tenant-scope grants (empty = access to all customers). Ignored for
            admins, who always access every customer.
          items: { type: string, format: uuid }
        lastLoginAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    AuthProviderType:
      type: string
      enum: [google, microsoft, github, oidc]
    AuthProviderConfig:
      type: object
      required: [id, type, name, displayName, clientId, enabled, source, redirectUri, createdAt]
      properties:
        id:
          type: string
          format: uuid
        type: { $ref: "#/components/schemas/AuthProviderType" }
        name: { type: string, description: "URL-safe slug used in /auth/{name}/start" }
        displayName: { type: string }
        clientId: { type: string }
        issuer: { type: string, nullable: true, description: Only for type oidc }
        enabled: { type: boolean }
        source:
          type: string
          enum: [database, environment]
          description: environment = defined via OTELFLEET_OIDC_* env vars (read-only here).
        redirectUri:
          type: string
          description: Callback URL to register at the identity provider.
        createdAt: { type: string, format: date-time }
    AuthProviderConfigCreate:
      type: object
      required: [type, name, displayName, clientId, clientSecret]
      properties:
        type: { $ref: "#/components/schemas/AuthProviderType" }
        name:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{1,30}$"
          description: Unique slug; becomes /auth/{name}/start.
        displayName: { type: string, minLength: 1, maxLength: 100 }
        clientId: { type: string, minLength: 1 }
        clientSecret: { type: string, minLength: 1 }
        issuer:
          type: string
          nullable: true
          description: Required for type oidc; ignored for google/microsoft/github.
        enabled: { type: boolean, default: true }
...
    AuditEntry:
      type: object
      required: [id, actorType, action, entityType, entityId, createdAt]
      properties:
        id: { type: integer, format: int64 }
        actorType: { type: string, enum: [user, system, agent] }
        actorUserId: { type: string, format: uuid, nullable: true }
        actorEmail: { type: string, nullable: true }
        action: { type: string, example: pipeline_version.activate }
        entityType: { type: string }
        entityId: { type: string }
        customerId: { type: string, format: uuid, nullable: true }
        customerName: { type: string, nullable: true }
        payload:
          type: object
          additionalProperties: true
          nullable: true
        createdAt: { type: string, format: date-time }
...